Demo reorganization, CSS Fixes for several components, Responsive fixed for Demo
