+++ src/components/demo/demoMenu/sliderInput/index.tsx
