+++ src/components/demo/demoMenu/sliderInput/slider.module.css
