+++ src/pages/demoLink/demoContent/demoContentLayout/index.css
@@ -1,4 +1,4 @@
-@media (max-height: 920px) {
+@media (max-height: 730px) {
     .demo-content {
         align-items: flex-start !important;
     }
