+++ src/pages/demoLink/demoContent/mobileContent/payCard/index.tsx
@@ -2,11 +2,33 @@ import { Box } from '@mui/material'
 import Typography from '@mui/material/Typography'
 import { getCssVar } from 'theme'
 import styles from './paycard.module.css'
+import './index.css'
 
 const PayCard = (): JSX.Element => {
   return (
     <Box
       className={styles.container}
+      id={'paycard-container'}
+      sx={
+        {
+          height: {
+            xs: '70px',
+            md: '120px'
+          },
+          minHeight: {
+            xs: '70px',
+            md: '120px'
+          },
+          width: {
+            xs: '120px',
+            md: '180px'
+          },
+          minWidth: {
+            xs: '120px',
+            md: '180px'
+          }
+        }
+      }
     >
       <Box
         className={styles['check-container']}>
@@ -34,9 +56,13 @@ const PayCard = (): JSX.Element => {
         }
       >
         <Typography
+          id={'paycard-text'}
           sx={
             {
-              fontSize: '14px',
+              fontSize: {
+                xs: '8px',
+                md: '14px'
+              },
               fontweigth: 'bold',
               color: getCssVar('--white')
             }
