+++ src/pages/demoLink/demoContent/mobileContent/payCard/paycard.module.css
@@ -1,8 +1,4 @@
 .container {
-    width: 180px;
-    height: 120px;
-    min-width: 180px;
-    min-height: 120px;
     background: linear-gradient(180deg, rgba(78, 163, 242, 0.85) 0%, rgba(65, 163, 255, 0.357) 100%);
     border-radius: 8px;
     padding: 5px;
