+++ src/pages/landing/whyUs/card/card.module.css
@@ -6,7 +6,7 @@
     margin: 0 8px;
     scroll-snap-align: center;
     position: relative;
-    box-shadow: 0px 55.128px 54.1025px rgba(49, 49, 49, 0.1);
+    box-shadow: 0px 45.128px 44.1025px rgba(49, 49, 49, 0.1);
 }
 
 .title {
